+++ cmd/get/nodepools/provider/common.go
@@ -1,13 +1,5 @@
 package provider
 
-import (
-	"strings"
-)
-
 const (
 	naValue = "n/a"
 )
-
-func formatCondition(condition string) string {
-	return strings.ToUpper(condition)
-}